requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.0.10
django-cors-headers==4.4.0
djangorestframework==3.15.1
idna==3.7
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
